+++ hc/front/forms.py
@@ -367,7 +367,7 @@ class GroupForm(forms.Form):
 
 class NtfyForm(forms.Form):
     error_css_class = "has-error"
-    topic = forms.CharField(max_length=50)
+    topic = forms.CharField(max_length=64)
     url = LaxURLField(max_length=1000)
     token = forms.CharField(max_length=100, required=False)
     priority = forms.IntegerField(initial=3, min_value=0, max_value=5)