+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 - Update CustomHeaderMiddleware to normalize email addresses to lower case (#1074)
 - Add data migration to convert existing user account emails to lower case
 - Update email alerts to mention failure reason (#1069)
+- De-emphasize the unsubscribe link in email notifications
 
 ### Bug Fixes
 - Improve recipient address validation in the smtp listener (#1077)